Fixes #124: Don't autoclose the popup when modifier keys are pressed
We originally had some problems with the popup staying open when the
user switched tabs or hit other special key combinations. However, now
that we are much less cavalier with preventing the popup from closing
(we only prevent it in the case that the popup is pinned), it seems that
we don't need to worry about this. Switching tabs, opening new windows,
entering the 'print screen' key combination, all seem to cleanly dismiss
the popup. The modifier key handling code is just not needed.