Fix #128: Convert inner fns to fat arrows & remove JS module globals
Several inner functions had scoping issues due to improper `this`
binding. Replace those functions with anonymous arrow functions, or pull
them out to named, top-level functions in cases where the function needs
to be unregistered. Use .bind in the bizarre case of a callback invoked
by XBL with a really weird global object (see code comments for all the
feels).

Convert module globals to instance properties. Module-global occurrences
of the `app` and `win` variables were not removed in the switch to the
Firefox module loader. Because the modules are cached after the first
load, those `app` and `win` globals were being overwritten with a
reference to the most-recently-opened window, breaking the popup in all
the other windows.

Add a `shutdown` method to Broker, fixing the proximate cause of #128.

These improvements also fix #129 and fix #130.